ProBowler-backend/requirements.txt
flask
flask-cors
flask-bcrypt
pymongo
dnspython
python-dotenv
mediapipe
opencv-python
numpy